+++ CHANGELOG.md
@@ -6,6 +6,7 @@ v0.12.3 (unreleased)
 * downgrade libpng to 1.2.53
 * **#1843**: add --keep-relative-links and --resolve-relative-links to resolve relative links
 * **#2104**: renamed COPYING to LICENSE
+* **#2183**: apply --javascript-delay settings for non-main resource loaders
 * **#2190**: do not depend on ICU even if it is already installed
 * **#2194**: **[qt]** fix debugging of Qt/WebKit due to missing debug information
 * **#2280**: do not allow data URIs for --header-html or --footer-html